Added number serialization mode  (#3)
A new serialization mode is added to allow users to serialize the quantity directly to its magnitude. This will drop the units in the serialization. The original implementation did not implementation due to this loss of information, but ultimately this is an "opt-in" feature, so it is more acceptable to implement this.